requirements.txt
dj-database-url==0.4.2
Django==1.11
Markdown==2.6.9
django-multiselectfield==0.1.12
django-phone-field
django-formwizard==1.0
python-decouple==3.1
django-dirtyfields==1.4
django-cors-headers==3.2.1
django-sslify==0.2.7
django-countries==6.0
django-crispy-forms==1.9.0
django-crum==0.7.5
django-tables2==2.2.1
django-widget-tweaks==1.4.5
django-datatable==0.3.1
gunicorn==19.9.0
psycopg2==2.7.7
redis==3.2.1
django-formtools==2.2
django-extensions==2.2.9
django-material==1.6.3
